requirements.txt
playwright
apscheduler
fastapi
uvicorn[standard]
python-dotenv
pandas
openpyxl
colorama
